+++ asset/src/id_reader/fetcher.ts
@@ -1,0 +1,4 @@
+
+import Processor from '../elasticsearch_reader/fetcher';
+
+export default Processor;
