+++ asset/src/id_reader/helpers.ts
@@ -1,0 +1,31 @@
+/* eslint-disable no-useless-escape */
+
+import { ESIDReaderConfig } from './interfaces';
+
+export const base64url = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w',
+    'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
+    'Y', 'Z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '\-', '_'];
+
+export const base64 = ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w',
+    'x', 'y', 'z', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X',
+    'Y', 'Z', '0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '\-', '_', '+', '/'];
+
+export const hexadecimal = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'a', 'b', 'c', 'd', 'e', 'f'];
+
+export const HEXADECIMAL = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F'];
+
+export function getKeyArray(config: ESIDReaderConfig) {
+    if (config.key_type === 'base64url') {
+        return base64url.slice();
+    }
+    if (config.key_type === 'base64') {
+        return base64.slice();
+    }
+    if (config.key_type === 'hexadecimal') {
+        return hexadecimal.slice();
+    }
+    if (config.key_type === 'HEXADECIMAL') {
+        return HEXADECIMAL.slice();
+    }
+    throw new Error('could not find correct key type');
+}
