+++ asset/src/simple_api_reader/interfaces.ts
@@ -1,0 +1,8 @@
+
+import { ESReaderConfig } from '../elasticsearch_reader/interfaces';
+
+export interface ApiConfig extends ESReaderConfig {
+    endpoint: string;
+    token: string;
+    timeout: number;
+}
